Add files via upload
Random Forest is a popular machine learning algorithm used for both classification and regression tasks. It is an ensemble learning method, which means it combines the predictions from multiple individual models to make more accurate and robust predictions. Random Forest is particularly well-suited for complex datasets and often outperforms single decision trees.

The Random Forest algorithm works:

Bootstrapping: Random Forest starts by creating multiple random subsets of the training data. Each subset is created by randomly sampling the data points with replacement. These subsets are known as "bootstrap samples."

Decision Trees: For each bootstrap sample, a decision tree is trained independently. These decision trees are often referred to as "base learners" or "weak learners." Each tree is trained on a different subset of the data and may split the data into branches based on various features.

Voting or Averaging: Once all the decision trees are trained, they can be used to make predictions. In classification tasks, each tree "votes" for a class, and the class with the most votes becomes the final prediction. In regression tasks, the predictions from all the trees are averaged to obtain the final prediction.

Randomness: Random Forest introduces randomness in two key ways:

Random feature selection: When building each tree, only a random subset of features is considered for each split. This helps prevent overfitting and decorrelates the trees. Bootstrap sampling: As mentioned earlier, each tree is trained on a random subset of the data with replacement. This leads to variation in the training data for each tree. The combination of multiple decision trees with these sources of randomness results in a robust and accurate ensemble model. Random Forests are less prone to overfitting compared to individual decision trees and are known for their ability to handle high-dimensional datasets with many features.

Some key advantages of Random Forest include:

Robustness against overfitting.

Good performance on a wide range of datasets.

Automatic feature importance estimation.

Ability to handle missing data and outliers.